Converted the "size" variable to an int.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@2789 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genometryImpl/src/com/affymetrix/genometryImpl/util/ByteList.java
@@ -28,7 +28,7 @@ public class ByteList {
   /**
    * The size of the *List (the number of primitive values it contains).
    */
-  private byte size;
+  private int size;
 
   public ByteList(int initialCapacity) {
     super();
